backend/requirements.txt
flask
openai
google-genai
gtts
flask-cors
